Melhoramento: - Inserção de informação na 2º linha - Criação de um método para popular a base de dados quando feito o registo do utilizador

+++ app/src/main/java/com/example/android/diadata/db/dao/MealDao.java
@@ -21,7 +21,7 @@ public interface MealDao {
 
     //Vai buscar todos os ingredientes da refeição e executa a soma
     @Query("SELECT SUM(qnt_hidratos) FROM meal WHERE nome_meal = :nome_refeicao")
-    int getSomaHidratos(String nome_refeicao);
+    double getSomaHidratos(String nome_refeicao);
 
     @Insert
     void addRefeicao(Meal meal);
